Fix bundle install error due to invalid homepage
```
> bundle install
Fetching https://github.com/sebastiangeiger/lita-identity-manager.git
You have one or more invalid gemspecs that need to be fixed.
The gemspec at
~/.rbenv/versions/2.6.5/lib/ruby/gems/2.6.0/bundler/gems/lita-identity-manager-0ce0d2e45ea6/lita-identity-manager.gemspec
is not valid. Please fix this gemspec.
The validation error was '"TODO: Add a homepage" is not a valid HTTP URI'
```

And in case it barfs on license the way that NPM likes to, pick one for that, too.